ruzhi/backend/ai-service/requirements.txt
fastapi>=0.85.0
uvicorn>=0.18.3
pydantic>=1.9.2
pydantic-settings>=2.0.0
python-dotenv>=0.21.0
sqlalchemy>=2.0.19
alembic>=1.11.1
psycopg2-binary>=2.9.6
python-jose>=3.3.0
python-multipart>=0.0.5
httpx>=0.24.1
loguru>=0.7.0
langchain>=0.0.267
sentence-transformers>=2.2.2
openai>=0.27.8
huggingface-hub>=0.16.4
pyjwt>=2.6.0
jinja2>=3.1.2
aiohttp>=3.8.3 